+++ doxtr_pdf_theme_core/ast_processors/__init__.py
@@ -12,6 +12,7 @@ Each processor handles a specific type of RST/Sphinx element:
 - highlights: Highlights directive styling
 - needs: sphinx-needs requirement boxes
 - topics: Topic and contents directive styled boxes
+- todo: sphinx.ext.todo directive styled boxes
 - landscape: Forced and automatic landscape page orientation
 """
 from .containers import process_containers_ast
@@ -26,6 +27,7 @@ from .sidebars import process_sidebar_ast, render_nodes_to_latex
 from .highlights import process_highlights_ast
 from .needs import process_needs_ast
 from .topics import process_topics_ast
+from .todo import process_todo_ast
 from .landscape import (
     process_landscape_ast,
     wrap_in_landscape,
@@ -56,6 +58,7 @@ __all__ = [
     'process_highlights_ast',
     'process_needs_ast',
     'process_topics_ast',
+    'process_todo_ast',
     'process_landscape_ast',
     'wrap_in_landscape',
     'register_landscape_wrapper',